chore(ui): fix renamed date utils file import in components
